% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_packages.R
\name{install_packages}
\alias{install_packages}
\title{Install packages}
\usage{
install_packages(re_install = FALSE)
}
\arguments{
\item{re_install}{Whether to re-install already installed packages (
default = FALSE)}
}
\value{
Nothing, but packages will be installed
}
\description{
Install some subjectively useful packages
}
