% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_colors.R
\name{make_colors}
\alias{make_colors}
\title{Make colors}
\usage{
make_colors(n = 3, categorical = TRUE, b2g = FALSE, r2g = FALSE)
}
\arguments{
\item{n}{How many colors}

\item{categorical}{Boolean, whether to make a categorical palette (alternative, numeric, assumes ordinality)}

\item{b2g}{Whether to use a blue to green palette (as opposed to blue to white);
only applicable if categorical is FALSE}

\item{r2g}{Whether to use a red to green palette (as opposed to blue to white); only applicable if categorical is FALSE}
}
\description{
Generate a color palette of length n
}
