% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_word_cloud.R
\name{prepare_for_word_cloud}
\alias{prepare_for_word_cloud}
\title{Prepare for word cloud}
\usage{
prepare_for_word_cloud(x)
}
\arguments{
\item{x}{A character vector}
}
\description{
Prepare a character vector for use with wordcloud2
}

