% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_sentiment.R
\name{score_sentiment}
\alias{score_sentiment}
\title{Score sentiment}
\usage{
score_sentiment(x)
}
\arguments{
\item{x}{A character vector of phrases to be scored}
}
\description{
Score the positivity / negativity of a statement
}

