% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streak.R
\name{streak}
\alias{streak}
\title{Identify streaks}
\usage{
streak(x, value = 0)
}
\arguments{
\item{x}{A vector}

\item{value}{The value to be counted}
}
\description{
Identify chained repetitions (streak) of a value in a vector, and count the length of the streak
}
