% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_databrew.R
\name{theme_databrew}
\alias{theme_databrew}
\title{databrew theme}
\usage{
theme_databrew(base_size = 18, base_family = "Roboto",
  subtitle_family = "Roboto", axis_family = "Roboto")
}
\arguments{
\item{base_size}{The size of the base font}

\item{base_family}{The font used in the legend and title}

\item{subtitle_family}{The font used in the subtitle}

\item{axis_family}{The font used in the axes}
}
\value{
A ggtheme object meant to be used in conjunction with a call to ggplot()
}
\description{
Apply the databrew look to a ggplot2-generated visualization
}
