% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_location.R
\name{visualize_location}
\alias{visualize_location}
\title{Visualize location}
\usage{
visualize_location(x, use_leaflet = FALSE, spdf = NULL)
}
\arguments{
\item{use_leaflet}{Whether to use leaflet}

\item{spdf}{A spatial polygons dataframe}

\item{data}{A dataframe}
}
\value{
A map
}
\description{
Visualize location of a dataframe with lat/lon columns
}
